User Scopes

+++ src/main/java/com/ntr1x/storage/uploads/services/ImageService.java
@@ -6,6 +6,7 @@ import java.util.UUID;
 import javax.imageio.ImageIO;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
+import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.ServerErrorException;
 import javax.ws.rs.core.Response;
 
@@ -21,6 +22,7 @@ import com.ntr1x.storage.core.services.IFileService;
 import com.ntr1x.storage.core.services.IScaleImageService;
 import com.ntr1x.storage.security.model.User;
 import com.ntr1x.storage.security.services.ISecurityService;
+import com.ntr1x.storage.security.services.IUserService;
 import com.ntr1x.storage.uploads.repository.ImageRepository;
 
 @Service
@@ -33,6 +35,9 @@ public class ImageService implements IImageService {
     private ImageRepository images;
     
     @Inject
+    private IUserService users;
+    
+    @Inject
     private IScaleImageService scale;
     
     @Inject
@@ -42,14 +47,15 @@ public class ImageService implements IImageService {
     private ISecurityService security;
     
     @Override
-    public Image upload(ImageCreate create) {
+    public Image upload(long scope, ImageCreate create) {
     	
     	Image image = new Image(); {
+    		image.setScope(scope);
             image.setUuid(UUID.randomUUID());
             image.setOriginal(create.original);
         };
         
-        User user = em.find(User.class, create.user);
+        User user = users.select(scope, create.user);
         
         em.persist(image);
         em.flush();
@@ -90,28 +96,29 @@ public class ImageService implements IImageService {
     }
     
     @Override
-    public Image remove(long id) {
+    public Image remove(Long scope, long id) {
     	
-    	Image image = em.find(Image.class, id);
-    	
-    	em.remove(image);
-    	em.flush();
+    	Image image = images.select(scope, id); {
+    		
+    		em.remove(image);
+        	em.flush();
+    	}
     	
     	return image;
     }
     
     @Override
-    public Image select(long id) {
-    	return images.findOne(id);
+    public Image select(Long scope, long id) {
+    	return images.select(scope, id);
     }
     
     @Override
-	public Image select(UUID uuid) {
-		return images.findByUuid(uuid);
+	public Image select(Long scope, UUID uuid) {
+		return images.select(scope, uuid);
 	}
     
     @Override
-    public Page<Image> query(String aspect, Pageable pageable) {
+    public Page<Image> query(Long scope, String aspect, Pageable pageable) {
     	
     	return aspect != null && !aspect.isEmpty()
             ? images.query(aspect, pageable)
@@ -130,6 +137,7 @@ public class ImageService implements IImageService {
                     
                     Image e = em.find(Image.class, p.image);
                     
+                    v.setScope(resource.getScope());
                     v.setRelate(resource);
                     v.setImage(e);
                     
@@ -154,8 +162,9 @@ public class ImageService implements IImageService {
                         
                         ResourceImage v = new ResourceImage(); {
                             
-                            Image e = em.find(Image.class, p.image);
-                            
+                            Image e = select(resource.getScope(), p.image);
+                    
+                            v.setScope(resource.getScope());
                             v.setRelate(resource);
                             v.setImage(e);
                             
@@ -169,8 +178,13 @@ public class ImageService implements IImageService {
                     }
                     case REMOVE: {
                         
-                        ResourceImage v = em.find(ResourceImage.class, p.id);
-                        em.remove(v);
+                        ResourceImage v = em.find(ResourceImage.class, p.id); {
+                        	
+                        	if (v.getRelate().getId() != resource.getId() || v.getRelate().getScope() != resource.getScope()) {
+                        		throw new ForbiddenException("Image relates to another scope or resource");
+                        	}
+                        	em.remove(v);
+                        }
                         break;
                     }
                 default:
